Move feature flag code to common-lib
