--- src/main/java/uk/gov/hmcts/opal/launchdarkly/FeatureDisabledException.java
@@ -1,9 +1,0 @@
-package uk.gov.hmcts.opal.launchdarkly;
-
-
-public class FeatureDisabledException extends RuntimeException {
-
-    public FeatureDisabledException(String message) {
-        super(message);
-    }
-}
