+++ src/test/java/uk/gov/hmcts/opal/controllers/advice/GlobalExceptionHandlerTest.java
@@ -65,7 +65,7 @@ import uk.gov.hmcts.opal.exception.JsonSchemaValidationException;
 import uk.gov.hmcts.opal.common.exception.OpalApiException;
 import uk.gov.hmcts.opal.exception.ResourceConflictException;
 import uk.gov.hmcts.opal.exception.UnprocessableException;
-import uk.gov.hmcts.opal.launchdarkly.FeatureDisabledException;
+import uk.gov.hmcts.opal.common.launchdarkly.FeatureDisabledException;
 
 @SpringBootTest
 @ContextConfiguration(classes = GlobalExceptionHandler.class)
