--- src/test/java/uk/gov/hmcts/opal/launchdarkly/FeatureDisabledExceptionTest.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.opal.launchdarkly;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class FeatureDisabledExceptionTest {
-
-    @Test
-    void testConstructor() {
-        // Arrange
-        String errorMessage = "Feature is disabled";
-
-        // Act
-        FeatureDisabledException exception = new FeatureDisabledException(errorMessage);
-
-        // Assert
-        assertEquals(errorMessage, exception.getMessage());
-    }
-}
