--- src/test/java/uk/gov/hmcts/opal/launchdarkly/FeatureToggleServiceTest.java
@@ -1,49 +1,0 @@
-package uk.gov.hmcts.opal.launchdarkly;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.when;
-
-@ExtendWith(MockitoExtension.class)
-class FeatureToggleServiceTest {
-
-    @Mock
-    private FeatureToggleApi featureToggleApi;
-
-    private FeatureToggleService featureToggleService;
-
-    @BeforeEach
-    void setUp() {
-        featureToggleService = new FeatureToggleService(featureToggleApi);
-    }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {true, false})
-    void shouldReturnCorrectValue_whenMyFeatureIsEnabledOrDisabled(Boolean toggleState) {
-        givenToggle("myFeature", toggleState);
-
-        assertThat(featureToggleService.isFeatureEnabled("myFeature")).isEqualTo(toggleState);
-    }
-
-    private void givenToggle(String feature, boolean state) {
-        when(featureToggleApi.isFeatureEnabled(eq(feature)))
-            .thenReturn(state);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = { "opal", "legacy"})
-    void shouldReturnCorrectValue_whenMyFeatureValueInvoked(String toggleValue) {
-        when(featureToggleApi.getFeatureValue(eq("myFeature"), any()))
-            .thenReturn(toggleValue);
-
-        assertThat(featureToggleService.getFeatureValue("myFeature")).isEqualTo(toggleValue);
-    }
-}
